+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   fun :return value)` or `(:mock fun :with other-fun)`.
 - `bydi` now accepts shorthand `(:always fun)` and `(:ignore fun)` to
   mock a function with `ignore` or `always` respectively.
+- `bydi` now accepts shorthand `(:toggle fun)` that will have a
+  function return `t` by default or `nil` if `(bydi-toggle-sometimes)`
+  was called.
 
 ## [v0.1.0]